1.视频信息的增删改查优化

+++ WYVideo/business/src/main/java/pxxy/wzf/business/rest/vo/param/VideoVO.java
@@ -21,7 +21,7 @@ public class VideoVO implements Serializable {
     /**
     * 总结id
     */
-    private Integer summaryId;
+    private Long summaryId;
 
     /**
     * 视频地址
